+++ CHANGELOG.md
@@ -6,9 +6,14 @@
 ### Added
 ### Changed
 ### Fixed
-- Changelog typo
-### Removed
 
+### Removed
+## [v.2.0.0] - 2019-01-23
+### 🔥 Changed (breaking changes)
+- Updated selectors to use new RouteContext approach introduced in PWA v6.0. **Breaking change**: PWA 6.x required
+- Updated `Link` component import. **Breaking change**: PWA 6.x required
+### Fixed
+- Changelog typo
 ## [v.1.0.2] - 2018-11-09
 ### Fixed
 - Modified Label component to account for location prop passed from product list view.
@@ -23,4 +28,6 @@
 First extension version.
 
 [Unreleased]: https://github.com/shopgate/ext-energy-efficiency-labels/compare/v.1.0.1...HEAD
+[v.2.0.0]: https://github.com/shopgate/ext-energy-efficiency-labels/compare/v.1.0.2...v.2.0.0
+[v.1.0.2]: https://github.com/shopgate/ext-energy-efficiency-labels/compare/v.1.0.1...v.1.0.2
 [v.1.0.1]: https://github.com/shopgate/ext-energy-efficiency-labels/compare/v.1.0.0...v.1.0.1